Move relevant files into converse-rootview plugin folder
