+++ src/plugins/chatboxviews/index.js
@@ -5,7 +5,6 @@
  */
 import './view.js';
 import '@converse/headless/plugins/chatboxes/index.js';
-import 'components/converse.js';
 import ChatBoxViews from './container.js';
 import { ViewWithAvatar } from 'shared/avatar.js';
 import { _converse, api, converse } from '@converse/headless/core';
