+++ src/plugins/rootview/templates/root.js
@@ -1,4 +1,4 @@
-import '../components/font-awesome.js';
+import 'components/font-awesome.js';
 import { api } from '@converse/headless/core';
 import { html } from 'lit-html';
